added details component, create nesting between tasks-task-details componets
